+++ PySubtitle/UnitTests/__init__.py
@@ -3,7 +3,7 @@ from PySubtitle.UnitTests.test_Subtitles import TestSubtitles, SubtitleProcessor
 from PySubtitle.UnitTests.test_Parse import TestParseDelayFromHeader, TestParseNames, TestParseValues
 from PySubtitle.UnitTests.test_Substitutions import TestSubstitutions
 from PySubtitle.UnitTests.test_Time import TestTimeHelpers
-from PySubtitle.UnitTests.test_Project import ChineseDinnerTests
+from PySubtitle.UnitTests.test_ChineseDinner import ChineseDinnerTests
 from PySubtitle.UnitTests.test_Translator import SubtitleTranslatorTests
 from PySubtitle.UnitTests.test_MergeSplitCommands import MergeSplitCommandsTests
 from PySubtitle.UnitTests.test_MergeLinesCommand import MergeLinesCommandTest
